// src/enums/screenNames.ts
export enum screenNames {
  HOME = "HOME",
  SPLASHSCREEN = "SPLASHSCREEN",
  LOGIN = "LOGIN",
  RESETPASSWORD = "RESETPASSWORD",
  OTPSCREEN = "OTPSCREEN",
  CONFIRMPASSWORD = "CONFIRMPASSWORD",
  DRAWERNAVIGATION = "DRAWERNAVIGATION",
  BOTTOMNAVIGATION = "BOTTOMNAVIGATION",
  EPISODELIST = "EPISODELIST",
  EPISODEDETAILS = "EPISODEDETAILS",
  TOCSLIST = "TOCSLIST",
  HOMENAVIGATION = "HOMENAVIGATION",
  EPISODENAVIGATION = "EPISODENAVIGATION",
  TOCNAVIGATION = "TOCNAVIGATION",
  CONTACT = "CONTACT",
  EX = "EX",
  TOCDETAILS = "TOCDETAILS",
  ALLCHATMESSAGES = "ALLCHATMESSAGES",
  CHAT = "CHAT",
  CREATECONVERSION = "CREATECONVERSION",
  NOTIFICATIONS = "NOTIFICATIONS",
  FACEIDTOUCHIDENABLE = "FACEIDTOUCHIDENABLE",
  PDFVIEWERSCREEN = "PDFVIEWERSCREEN",
  ABOUTENAVPROVIDER = "ABOUTENAVPROVIDER",
}